Seperated speaker and speakercontroller
